// src/pages/createJob/changeCorporateAccount/edit/constants/types.ts
export type TSuitTest = {
  id: string;
  questionNumber: number;
  question: string;
  choices: TSuitChoice[];
  types: string;
};

export type TSuitChoice = {
  id: string;
  questionId: string;
  answer: string;
  score: number;
};
export type TSuitAns = {
  id: string;
  quiz: number;
  ans: number | number[];
  type: number;
};
// ...
export type SuitTestResult = {
  corporateCode?: string;
  totalScore?: number;
  level?: number;
  invsetorTypeRisk?: string;
  suitTestResult?: SuitTestResultAnswer;
  type?: number;
};

export type SuitTestResultAnswer = {
  answer: SuitAnswer[];
  additional: Array<boolean | undefined | null>;
};

export type SuitAnswer = {
  id?: string;
  ans: number[] | number;
  type: number;
  quiz: number;
};